PWN-4397 - Add bottom navigation component
