+++ ui-kit/src/main/res/drawable/ic_feedback_main.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+
+    <path
+        android:fillColor="#6F7D8D"
+        android:fillType="evenOdd"
+        android:pathData="M2.03091 8.13349C2 8.73757 2 9.47286 2 10.4V13.6C2 15.8402 2 16.9603 2.43597 17.816C2.81947 18.5686 3.43139 19.1805 4.18404 19.564C5.03969 20 6.15979 20 8.4 20H15.6C17.8402 20 18.9603 20 19.816 19.564C20.5686 19.1805 21.1805 18.5686 21.564 17.816C22 16.9603 22 15.8402 22 13.6V10.4C22 9.47286 22 8.73757 21.9691 8.13349L13.7889 12.2236C12.6627 12.7867 11.3373 12.7867 10.2111 12.2236L2.03091 8.13349ZM21.5292 6.11734C21.1441 5.39524 20.5463 4.80808 19.816 4.43597C18.9603 4 17.8402 4 15.6 4H8.4C6.15979 4 5.03969 4 4.18404 4.43597C3.45375 4.80808 2.85595 5.39524 2.47075 6.11734L11.1056 10.4348C11.6686 10.7163 12.3314 10.7163 12.8944 10.4348L21.5292 6.11734Z" />
+</vector>
