+++ ui-kit/src/main/res/drawable/ic_history_main.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+
+    <path
+        android:fillColor="#6F7D8D"
+        android:fillType="evenOdd"
+        android:pathData="M12 22C17.5228 22 22 17.5228 22 12C22 6.47715 17.5228 2 12 2C6.47715 2 2 6.47715 2 12C2 17.5228 6.47715 22 12 22ZM13 7C13 6.44772 12.5523 6 12 6C11.4477 6 11 6.44772 11 7V12.4854C11 13.1313 11.3119 13.7374 11.8375 14.1128L14.9188 16.3137C15.3682 16.6347 15.9927 16.5307 16.3137 16.0812C16.6347 15.6318 16.5307 15.0073 16.0812 14.6863L13 12.4854V7Z" />
+</vector>
