+++ ui-kit/src/main/res/drawable/menu_select_deselect.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:color="#6F7D8D" android:state_selected="false" />
+    <item android:color="#0F0F0F" android:state_selected="true" />
+</selector>
